+++ src/config/share_target.rs
@@ -1,6 +1,12 @@
-use std::{fmt::Display, str::FromStr};
+use std::{
+    fmt::Display,
+    process::{Command, Stdio},
+    str::FromStr,
+};
 
 use crate::prelude::*;
+use arboard::Clipboard;
+use log::info;
 use percent_encoding::{NON_ALPHANUMERIC, utf8_percent_encode};
 use strum::EnumMessage;
 use url::Url;
@@ -46,12 +52,16 @@ pub enum ShareTarget {
     Telegram,
     #[strum(message = "Custom", detailed_message = "custom target")]
     Custom(String, String),
+
+    #[strum(message = "Command", detailed_message = "command")]
+    Command(String, Vec<String>),
 }
 
 impl AsRef<str> for ShareTarget {
     fn as_ref(&self) -> &str {
         match self {
             ShareTarget::Custom(name, ..) => name,
+            ShareTarget::Command(name, ..) => name,
             target => target.into(),
         }
     }
@@ -62,6 +72,7 @@ impl Display for ShareTarget {
         use ShareTarget as T;
         match self {
             T::Custom(name, ..) => f.write_str(name),
+            T::Command(name, ..) => f.write_str(name),
             target => f.write_str(target.get_message().unwrap_or(self.as_ref())),
         }
     }
@@ -79,7 +90,13 @@ impl FromStr for ShareTarget {
             "mastodon" => T::Mastodon,
             "telegram" => T::Telegram,
             s => match s.split_once(' ') {
-                Some((name, url_template)) => T::Custom(name.to_owned(), url_template.to_owned()),
+                Some((name, rest)) => {
+                    if rest.starts_with("http://") || rest.starts_with("https://") {
+                        T::Custom(name.to_owned(), rest.to_owned())
+                    } else {
+                        T::Command(name.to_owned(), shell_words::split(rest)?)
+                    }
+                }
                 None => {
                     return Err(ConfigError::ShareTargetParseError(format!(
                         "unable to parse share target: {s}"
@@ -103,6 +120,7 @@ impl ShareTarget {
             T::Telegram => TELEGRAM_URL_TEMPLATE,
             T::Custom(_, url_template) => url_template,
             T::Clipboard => return Err(ConfigError::ShareTargetInvalid),
+            T::Command(..) => return Err(ConfigError::ShareTargetInvalid),
         };
 
         let url_str = url_template
@@ -110,6 +128,53 @@ impl ShareTarget {
             .replace("{url}", &url_escaped)
             .replace("{title}", &title_escaped);
         Ok(Url::parse(&url_str)?)
+    }
+
+    pub fn share(&self, title: &str, url: &Url) -> color_eyre::Result<()> {
+        match self {
+            ShareTarget::Clipboard => {
+                let mut clipboard = Clipboard::new()?;
+                clipboard.set_text(url.to_string())?;
+                Ok(())
+            }
+
+            ShareTarget::Command(_, args) => self.execute_as_command(args, title, url),
+
+            target => {
+                let share_url = target.to_url(title, url)?;
+                webbrowser::open(share_url.to_string().as_str())?;
+                Ok(())
+            }
+        }
+    }
+
+    pub fn execute_as_command(
+        &self,
+        command_args: &[String],
+        title: &str,
+        url: &Url,
+    ) -> color_eyre::Result<()> {
+        let Some((cmd, args)) = command_args.split_first() else {
+            return Err(color_eyre::eyre::eyre!("command is empty"));
+        };
+
+        let quoted_args = args
+            .iter()
+            .map(|arg| {
+                arg.replace("{url}", shell_words::quote(url.as_ref()).as_ref())
+                    .replace("{title}", shell_words::quote(title).as_ref())
+            })
+            .collect::<Vec<String>>();
+
+        info!("executing command {} {:?}", cmd, quoted_args);
+        let _child = Command::new(cmd)
+            .stdin(Stdio::null())
+            .stdout(Stdio::null())
+            .stderr(Stdio::null())
+            .args(quoted_args)
+            .spawn()?;
+
+        Ok(())
     }
 }